fix: resolve all ruff E701 lint errors and syntax issues in test files
- Split test_agent_scorecard.py into two files (was concatenated)
- Fix unclosed string literal in test_pii_redactor.py
- Fix E701 multi-statement if/elif lines in test_auto_close_service.py
- Fix E701 for-loop one-liners in test_duplicate_cache_limit.py
- Fix E701 sys.modules stubs in test_supabase_service.py
- Fix E701 sys.modules stubs in test_tenant_isolation.py
- Fix special characters in docstrings (>= replaces unicode >=)
- Add missing get_encryption_key() to encryption.py
- Fix test_translation_service.py import ordering
- Update pytest.ini norecursedirs to exclude .tmp-ci-venv

:wq#
