+++ pytest.ini
@@ -1,2 +1,3 @@
 [pytest]
-pythonpath = . backend
+testpaths = backend/tests
+norecursedirs = .tmp-ci-venv .venv venv env node_modules __pycache__ .git
